Move sidebar with upsell cards to shared space
