+++ packages/js/src/shared-admin/components/index.js
@@ -1,0 +1,4 @@
+export { AcademyUpsellCard } from "./academy-upsell-card";
+export { ReactComponent as G2LogoWhite } from "./g2-logo-white.svg";
+export { PremiumUpsellCard } from "./premium-upsell-card";
+export { RecommendationsSidebar } from "./recommendations-sidebar";
